Clean up activity on pause (force close) and fix issues with media player
Now finishing the activity no matter what when it is paused. Also I
changed the way the media player works - if you press a button for a
sound while a sound is playing, the playing sound will stop now and the
new sound will start playing. Multiple sounds do not play at once. This
seems to have solved the issue where none of the buttons work suddenly.